requirements.txt
ipython
pandas
numpy
matplotlib
seaborn
flask
dagshub
mlflow <3
scikit-learn
imbalanced-learn
dvc
pyyaml
dill
google-cloud-storage
lightgbm
xgboost
fastapi
Jinja2
uvicorn
ipykernel
python_box
beautifulsoup4
python_dotenv
geopandas
statsmodels
plotly
ensure

#-e .
ipython
pandas
numpy
matplotlib
seaborn
flask
dagshub
mlflow <3
scikit-learn
imbalanced-learn
dvc
pyyaml
dill
google-cloud-storage
lightgbm
xgboost
fastapi
Jinja2
uvicorn
ipykernel
python_box



#-e .